Fix match type size (8 vs 16 bits).
Filtering the match display only used 8 of the 16 bits.

Commit 071bfbfdcc in Feb 2023 changed the size of match_type, but missed
updating the cast in match_display_filter_entry.

Now match_display_filter_entry uses the real match_type.

+++ clink/lua/src/lua_match_generator.cpp
@@ -399,7 +399,7 @@ done:
                 match_display_filter_entry *new_match;
                 new_match = (match_display_filter_entry *)malloc(alloc_size);
                 memset(new_match, 0, sizeof(*new_match));
-                new_match->type = uint8(type);
+                new_match->type = type;
                 new_match->append_char = append_char;
                 new_match->flags = flags;
                 new_matches[j] = new_match;
